+++ src/modules/auth/auth.module.ts
@@ -6,16 +6,19 @@ import { AuthController } from '@/modules/auth/auth.controller'
 import { AuthService } from '@/modules/auth/auth.service'
 import { AuthUtilsService } from '@/modules/auth/auth.utils.service'
 import { CaptchaService } from '@/modules/auth/captcha.service'
+import { InternalAuthController } from '@/modules/auth/internal-auth.controller'
+import { InternalAuthGuard } from '@/modules/auth/internal-auth.guard'
 import { PasswordService } from '@/modules/auth/password.service'
 
 @Module({
     imports: [TypeOrmModule.forFeature([TbAccountUser])],
-    controllers: [AuthController],
+    controllers: [AuthController, InternalAuthController],
     providers: [
         AuthService,
         AuthUtilsService,
         AuthSessionService,
         CaptchaService,
+        InternalAuthGuard,
         JwtAuthGuard,
         PasswordService,
         TokenService,